% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_slope_analysis.R
\name{meta_slope_analysis}
\alias{meta_slope_analysis}
\title{Meta-analysis of standardized slopes (Cohen's d) from multiple studies}
\usage{
meta_slope_analysis(cov_matrices, mean_vectors, sample_sizes)
}
\arguments{
\item{cov_matrices}{A list of covariance matrices}

\item{mean_vectors}{A list of mean vectors}

\item{sample_sizes}{A vector of sample sizes}
}
\value{
A list with meta-analytic estimate of Cohen's d, confidence interval, and p-value
}
\description{
Meta-analysis of standardized slopes (Cohen's d) from multiple studies
}
\examples{
# See README for a reproducible example
}
